requirements.txt
math
matplotlib
numpy
operator
os
pytest
random
scipy.stats
subprocess


